+++ example/android/app/src/main/kotlin/software/handmade/flutter/flutter_siri_suggestions_example/MainActivity.kt
@@ -1,0 +1,6 @@
+package software.handmade.flutter.flutter_siri_suggestions_example
+
+import io.flutter.embedding.android.FlutterActivity
+
+class MainActivity: FlutterActivity() {
+}